[NFC] [DTrans] Move functions related to managing the CallInfo objects out of the DTransAnalysisInfo class to separate manager class
Move the functions for CallInfo object management to a separate class,
CallInfoManager, that can be used by both DTransAnalysisInfo and DTransSafetyInfo.
Move the declaration to in DTrans.h
Move the implementation to DTransUtils.cpp
Keeps the interface for DTransAnalysisInfo the same for the rest of DTrans to
make this transparent to the analysis and transformations.

Change-Id: Icd452eacb7bf4fd662a603b46b525b5e37a90fd2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/320138
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>