Implement CL_CONFIG_CPU_FORCE_WORK_GROUP_SIZE
Force CPU to work with specified workgroup size. If this env is set,
local_work_size of clEnqueueNDRangeKernel is ignored.

Value of the env can be a size or array of size. If array size is larger
than work_dim of clEnqueueNDRangeKernel, only the first work_dim values
of the array are used. If array size is smaller than work_dim, workgroup
size of higher dim will be set to 1.

Change-Id: I853b25761c8ccddadb7aab2282d46f85f640215e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/319020
Reviewed-by: Wenju He <wenju.he@intel.com>
Tested-by: Wenju He <wenju.he@intel.com>