[Driver] Add support for allowing Linux options for clang-cl
When using clang-cl --dpcpp (DPCP++ mode) we allow for the usage of
Linux based options when the /Q_allow-linux command line is used.
This is not currently enabled with 'dpcpp'.

Also perform some cleanup for checking for DPC++ mode to use the
new variable setting instead of checking for the command line option
Partially Addresses:  CMPLRLLVM-25356

Change-Id: Ie5aae474e2e74bb27ac9744981d18d1b9338b661
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/321273
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>