CMPLRLLVM-27596: Include the attribute "mustprogress" in the libfuncs
This change set adds the attribute "mustprogress" as part of the
libfuncs. The attribute basically means that a function or a loop
will never stall in an infinite loop. This attribute is needed
because the community now assumes that a function will never return
unless "willreturn" or "mustprogress" attributes are added to the
function. The attribute "willreturn" implies "mustprogress".
This thread contains some information about the issue:

https://lists.llvm.org/pipermail/llvm-dev/2021-January/148047.html

The main problem for us is that dead code elimination now won't
perform on calls that aren't marked as "willreturn"
(https://reviews.llvm.org/D96993). In the case of 520.omnetpp, the
libfunc dynamic_cast was marked as nothrow (which was incorrect),
but not "willreturn". We cannot mark it as "willreturn" because the
definition of "willreturn" means that there is no call to exit or
exception handling inside the function and will return to the previous
point in the stack. The libfunc dynamic_cast can call to the
exception (std::bad_cast). Therefore we need to use "mustprogress".

This change set restores 3.4% of performance for 520.omnetpp.

Change-Id: Ifc64dabdc9f79ab1310da5017934bfec1ce49d40
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/321403
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>