[OpenCL] Fix wlATIMonteCarlo.ll LIT test failure
Update CHECK-NOT for starting with header label in wlATIMonteCarlo.ll (a
vectorizer LIT test), aligned with other LIT tests (like wlATIReduction.ll).
preheader label is emitted by loop-simplify pass for for-loop which leads to LIT
test failure.
Also deleted the comments in CHECK-NOT with footer label, since this
check is too stricter and may be miss some failures.

Change-Id: I03adf35ffba78813a8ee1624d36ddbccf8e48ef7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/321322
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Tested-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>