[Driver][NFC] Cleanup some option setting for SYCL offload (#3542)
The area in which we setup options for the device and host compilations
is a bit disorganized.  Clean up some of this to be a little easier to
find where additional options should go.  Also clean up the convoluted
usage of UseSYCLTriple which was used largely in part when doing the old
method of static device lib compilations.

Change-Id: I6519b2a13cd2b55d6cc22e10ff7a981320865e5f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/321910
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Reviewed-by: Chupin, Pavel V <pavel.v.chupin@intel.com>
Tested-by: Chupin, Pavel V <pavel.v.chupin@intel.com>