tmp alloca bypass: ensure that the replacement dominates all alloca uses
After 077bff39d46364035a5dcfa32fc69910ad0975d0,
isDereferenceableForAllocaSize() can recurse into selects,
which is causing a problem for the new test case,
reduced from https://lists.llvm.org/pipermail/llvm-commits/Week-of-Mon-20210412/904154.html
because the replacement (the select) is defined after the first use
of an alloca, so we'd end up with a verifier error.

Now, this new check is too restrictive.
We likely can handle *some* cases, by trying to sink all uses of an alloca
to after the the def.

Disable memcpy of alloca bypass for instruction sources

This transformation is fundamentally broken when it comes to dominance,
it just happened to work when the source of the memcpy can be moved into
the place of the alloca. The bug shows up a lot more often since
077bff39d46364035a5dcfa32fc69910ad0975d0 allows the source to be a
switch.

It would be possible to check dominance of the source and all its
operands, but that seems very heavy for instcombine.

Change-Id: I6adf336a189338a24ca04594554a82af1e1b0f95
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/322329
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Tested-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
(cherry picked from commit d16b2cd2cd7b510199a409b1c6b28f3d52374e59)
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/322430