Add an implicit taskgroup around taskloop constructs.
Emit the implicit taskgroup directive around taskloop directives unless
nogroup is used in user code. The outer taskgroup directive should have
an auxiliary clause QUAL.OMP.IMPLICIT, which will let Paropt know that
the directive was implicitly generated by the frontend, and doesn't come
from user code.
If there is a {[reduction}} clause on a taskloop in the user code,
break it into:
task_reduction clause on the outer implicit taskgroup directive.
in_reduction clause on the taskloop directive.

The effect of any {{in_reduction}} clause is as if a reduction clause
with the same reduction operator and list items is present on the
simd construct.

Implements cmplrllvm-26464

Change-Id: I529c88d31d0c8ced17d295ab2191374b9164de4f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/321479
Tested-by: Yu, Jennifer <jennifer.yu@intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>