[1/2][NFC] Add test to reveal a bug in VPlan LLVM-IR vector CG handling of sincos.
VPlan LLVM-IR vector codegen does special handling of glibc
sincos calls in-order to be compliant with the SVML sincos
signature. However this specialization should not be applied
if scalar sincos is being vectorized via vector-variant. This
patch adds a test to reveal a bug in CG when sincos is lowered
via vector-variants instead of SVML library call.

JIRA : CMPLRLLVM-27551

Change-Id: I8f008dc4ea6641b7153ce36dd027b9a99c8946f4
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/321444
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>