[2/2][VPlan] Don't reduce number of arg-operands for vector-variant based sincos.
VPlan's LLVM-IR vector codegen does special handling of glibc
sincos to be compliant with vectorized __svml_sincos function.
As part of this specialization, number of argument operands of
generated vector call should be reduced by 2, as sin/cos values
are not returned by pointers in __svml_sincos. However this
should not be done if sincos is going to be vectorized via
vector-variants instead of SVML library call. This patch fixes
the above bug in VPlan CG.

JIRA : CMPLRLLVM-27551

Change-Id: Ief8899ec035052cf96f4c60b4e231239627a00e4
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/321445
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>