CMPLRLLVM-10459: Global dope vectors constant propagation (part 1)
This change set enables the collection of constant information for
global variables that are dope vectors. It will traverse through
each global variable and identify where each field of the dope vector
is allocated, loaded and stored. Once this information is collected
then it validates that each field was accessed correctly and tries
to identify the possible constant value.

Change-Id: Id7edbf177369a177e034c321c63a14a47d508017
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/322209
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>