[Driver][SYCL][FPGA] Use opencl-aot for AOT when doing FPGA emulation (#3440)
When performing AOT compilation for FPGA, we call 'aoc' to perform
the offline compilation. Update this behavior to use opencl-aot when
we know we are performing emulation instead of running on hardware.

When creating FPGA AOCX/AOCR archives when performing emulation, create
archives which contain .spv items as opposed to using the regular aocx
and aocr binaries.

The expected input type for AOCX/AOCR were considered for hardware
usage with FPGA. In this part, we allow for the differentiation of
the type of archive created, whether it be emulation or hardware.
Depending on how these archives are created, we make sure we take
the proper flow to process the contents.

Also perform some cleanup in the needed triple arch representation
for fpga, where we didn't need to define all of the different types
we were wrapping for FPGA specific binaries.

This is a cherry-pick of:  https://github.com/intel/llvm/commit/3a995586491ef9da383ff71d0599a8e20b3c5b07

Change-Id: Ic9afcec405e1a8a1d898f840b9eb9c8d520ad719
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/322194
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>