[Paropt] Change MemberOf map-type update code to honor the value sent by FE.
Previously, Paropt was making the memberof flag point to the base of the
map chain unconditionally.

Now, Paropt honors the value sent by the frontend, and if it's pointing
to a link in the middle of the chain, then the update takes that into
account.

There's a flag to revert to the previous behavior:
  -vpo-paropt-force-member-of-to-point-to-base=true

For ifx, the previous behaviur is kept as the FFE always uses "1" as
the value of member-of. This exclusion can be removed once FFE fixes
CMPLRLLVM-27688.

Change-Id: I83c1ec91e23ecb6fa4ef52ce219111627fb7891b
JIRA: CMPLRLLVM-27605
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/321667
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>