CMPLRLLVM-24670: Re-enable this test for Windows
The problem was that the legacy call graph SCC pass manager uses an
on-the-fly version of Tarjan's algorithm to compute SCCs, which are
then fed to the SCC passes. The SCC iterator class used for this
contains a VisitStack with iterators to call graph nodes. If an optimization
does dead function elimination, this can eliminate call graph nodes to
which there are iterators on the stack, and dereferencing those iterators
will cause a core dump.

The community solved the problem by introducing an additional argument to
runAttributorOnFunctions() in Attributor.cpp, which is set to false for
the legacy pass manager version so that it does not delete dead functions.

Therefore, it is safe to re-enable this test.

Change-Id: If07abb7b808fc253e4c2c8b771c63892d416fb23
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/321877
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>