Start fixing CMPLRLLVM-27372.
There are numerous problems to be fixed, so this is an incremental
improvement more than a full fix. More work is needed to fix this
properly.

No longer reuse the annotation token for loop_hint when the
preprocessor parses a non-loop_hint pragma. When the parser
processes the new annotation token, it can now pass the correct
number of arguments to the parsed attribute being created.

Added a massive hack to the attr tablegen emitter to work around
the fact that the LoopHint attribute specifies all of the pragma
spellings in Attr.td, but that a different semantic attribute than
LoopHint is what's actually created. Without this hack, diagnostics
involving 'ivdep' would assert.

Change-Id: I22263b7ed59f8db4877fbcdf1b6c830162ec625e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/322110
Reviewed-by: Chupin, Pavel V <pavel.v.chupin@intel.com>
Tested-by: Chupin, Pavel V <pavel.v.chupin@intel.com>