CMPLRLLVM-27715: Bail out of createNodeForIdenticalOperandsPHI() for unreachable blocks.
This intel customization is causing changes in caching/querying behavior of SCEV
resulting in changes in range info causing lit test failures.

It is not worth processing phis with unreachable predecessors. The customization
wasn't intending to target such phis. It is better to bail out to avoid lit test
failures.

Change-Id: I544ccd936a8e590151cfce4ba65250d85a6877c2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/322154
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>