[DTrans - OpaquePtr] Cache information about safe pointer subtract instructions for the Transformations
The transformations need to adjust multiplier/divisor values associated with the
structure size when computations are done to compute the distance between two
pointers associated with an array of structures. This change saves the information
with the DTrans safety info so that it does not need to be recomputed after the
safety analysis has been run.

Change-Id: I0c3b4d273bf045116c90199f1212d73b47c84d78
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/322159
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>