[NFC][VPlan] Don't use insignificant address bits used to print names in Vplan.
I've got an issue during testing of cfg merger, when the names of instructions from
different vplans were printed identically, see VP0 usage in
vplan_new_cfgmerge_m4_v4_m4s1.ll.
The suggested solution is to use bits [4:19] of the address when we print names.
Currently we use [0:15], lower four of which are 0 always.

Change-Id: I17c76fa04f9c5cdf7a7c77118f97758080e1112b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/322392
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>