CMPLRLLVM-27805: Fix for regression in 511.povray and partial inlining
The current implementation of the simplified partial inlining looks
for one argument that will be used to access the exit block. There
is a chance that multiple arguments can branch into the exit block.
The analysis process can choose the wrong argument to check if
it is used to compute a loop, which could produce a false result.
This change set collects all the arguments that could branch into
and exit block and then identify which one will be used for
computing the loop.

Change-Id: I7a4238d04930e5984bececf28998d45ecaa0f998
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/322599
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>