[CMPLRLLVM-27761] [Transpose] Fix compfail from Transpose analysis and transformation not handling subscript result in PHINode
The Transpose processing of subscript instructions was expecting that a
subscript result that was not the final dimension would always be directly
used as a parameter to the subscript call for the next dimension when
performing the profitability and stride transformation. This fixes a
problem when the subscript result instead gets used by a PHINode or Select
instruction by extending the code to traverse the users of the PHI/Select
to get to the next subscript call.

Change-Id: I87b400094ad167d46f2cd91a95aab1033b6c62c5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/322680
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>