[NFC] CMPLRLLVM-27465: Add whole program utils as part of LTO config
Currently, the whole program utils (WPUtils) is global variable in
the support directory. This variable is used to communicate between
the linker and the compiler. The issue is that the sanitizer build
found that it produces an use of unitializer value issue.

This change set fixes the problem by making WPUtils as a field of
the LTO Config structure. The linker (BFD/GOLD and LLD) use this
structure to set all the configuration needed for LTO. The WPUtils
will now be setup in the config and then it will be passed to the
whole program analysis. The compiler can access the WPUtils by
calling the function getWholeProgramLinkerUtils from the whole
program analysis (WPInfo).

Change-Id: I38a7c5bd5f88b6718c799a5bb8f86481320bf576
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/323143
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>