Cherry pick https://github.com/intel/llvm/pull/3612 to xmain-rel
[SYCL][PI][L0] Submit open command batch on event status query (#3612)

This closes and executes a queue's open command batch list when an
event that was created in the queue has its execution status
queried. This isn't strictly required, but if the execution status
is being queried, it seems like a good idea to make sure any existing
commands have been submitted. This fixes a situation where if you query
over and over on the event's status without ever enqueueing more
commands, or doing a wait, the command may remain unsubmitted.

Change-Id: Ib3c6e62dc3043150a40704e405c9cac8c2281914
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/323618
Reviewed-by: Kevin B. Smith <kevin.b.smith@intel.com>
Tested-by: Kevin B. Smith <kevin.b.smith@intel.com>