[DTrans] Fix deletefield26.ll LIT test to have meaningful CHECK lines for intended behavior of test
The CHECK lines in this test were never fully enabled after the functionality
was implemented. The transformation was being inhibited because all the fields
would have been deleted due to the identify unused field values functionality,
which allowed the test to pass the check that indicated the structure was unchanged.

Change-Id: I5bb9e39d6028b0e983ab467f20dd1941bb25a4a8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/323662
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>