[InjectTLIMapping] Don't expect vector intrinsic while verifying builtin function prototypes.
Before creating TLI mappings for a vectorizable function call we
perform an additional verification of the called function's
prototype in order to catch incompatible redeclaration of builtin
functions (https://git-amr-2.devtools.intel.com/gerrit/#/c/235166/).
This check is performed for non-intrinsic calls and we don't need
a vector-version of the intrinsic to do this check. This patch
relaxes this expectation which in-turn allows TLI mappings to be
emitted for functions like @llvm.ldexp which don't have valid
vector intrinsics.

JIRA : CMPLRLLVM-27799, CMPLRLLVM-27743

Change-Id: I39d0baa2e2612291269bea7507ad09b6e8b8f431
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/323861
Reviewed-by: Alexey Bataev <alexey.bataev@intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>