CMPLRLLVM-28054: Create correct PHI node for many loops specialization clone with return
While looking through today's list of Klocwork issues, I didn't find anything real
from Klocwork, but I did note an error in the code in the vicinity of one of the
Klocwork messages. A PHI node was being created with only one input. This fixes that
and adds a pair of LIT test to catch the error.

Change-Id: Ie72a75518d080d15655fa944e49d6193f2e388d8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/324271
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>