CMPLRLLVM-10459: Global dope vector constant propagation (part2)
This change set introduces the data collection for nested dope
vectors from the global dope vectors. The key idea is to find
where the global array is loaded and trace the array entry.
Each array entry is a structure where the fields are dope
vectors too. Once the nested dope vector is found, then we
proceed to collect where each field is used for loading and
storing data.

Change-Id: I621ac68695e939b924a11fa28e8f8d597f151402
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/324357
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>