CMPLRLLVM-28054: Create correct PHI node for many loops specialization clone with return
While looking through today's list of Klocwork issues, I didn't find anything real
from Klocwork, but I did note an error in the code in the vicinity of one of the
Klocwork messages. A PHI node was being created with only one input. This fixes that
and adds a pair of LIT test to catch the error.

Change-Id: I6c7dbbc4943fd34b42feadd3e743013285ba9ed1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/324649
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Cox, Robert <robert.cox@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>