CMPLRLLVM-27465: Add whole program utils as part of LTO config
Currently, the whole program utils (WPUtils) is global variable in
the support directory. This variable is used to communicate between
the linker and the compiler. The issue is that the sanitizer build
found that it produces an use of unitializer value issue.

This change set fixes the problem by making WPUtils as a field of
the LTO Config structure. The linker (BFD/GOLD and LLD) use this
structure to set all the configuration needed for LTO. The WPUtils
will now be setup in the config and then it will be passed to the
whole program analysis. The compiler can access the WPUtils by
calling the function getWholeProgramLinkerUtils from the whole
program analysis (WPInfo).

This change set also contains the fixes for reorganizing the
pass manager to perform whole program analysis at the beginning.
This secures that the whole program utils were set before any use.

Change-Id: I14344db1d1a5ea06e7f098b8241bb100e17b84e0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/324702
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>