CMPLRLLVM-27465: Reorganize WP analysis pass in the pass manager
This change set moves the whole program analysis before the
GlobalDCE pass in the pass manager. The issue is that GlobalDCE
calls the whole program analysis, if WPUtils is not set at
this point then it could produce incorrect results until
WPUtils is set. In simple words, the WP analysis is the first
thing we should do in the pass manager.

Change-Id: Ideabd15664319cb6934921b7eb4883838eebbe0e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/324733
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>