[VPlan][InstructionCost] Accept TTI's InstructionCost but still convert to unsigned
This commit clean ups the handling of the TTI's InstructionCost returned values
by using a thin wrapper defined inside VPlanTTIWrapper with the same name. This
was we prepare the code for straight llvm::InstructionCost propagation in
future, yet maintain the old processing inside VPlan.

We don't implement the changes thoughout the whole VPlan pipeline right now
because it's unclear how llvm::InstructionCost will handle the separation
between InfiniteCost and UnknownCost that might need to behave differently.

Change-Id: I40e19f0cbddd184cefc4cc7dec29ee57822a8989
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/324875
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>
Reviewed-by: Artem Gindinson <artem.gindinson@intel.com>