[CMPLRLLVM-28174] [Devirt] Make devirtualization vars be named.
When the devirtualization pass runs it may create a new global
constant for an existing variable, and then take the name of
variable for an Alias object, leaving the global constant
unnamed. This creates a problem for IR dumps that created after
the devirtualization, such as when using the -dump-module-before-dtrans
option to debug DTrans passes. The IR will have unnamed
global variables which will crash the whole-program-assume pass
because that triggers code that uses the getGUID function which
requires variables to be named. This change supplies a name value
for the new globals created.

Change-Id: I125882853826bbab212e7e6b839f06006d52ca26
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/325518
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>