[Paropt] Change insertion point of offload array allocas for target maps.
Previously, for constructs which involve maps, the allocas for the
offload_baseptr/ptr/mapper arrays were being inserted immediately before the
construct's entry directive. They are now inserted in the beinning of the
innermost parent region that would be outlined, if any, otherwise in the
entry block of the parent function.

The same change is done for allocas "%.run_host_version" and
"%promoted.clause.args" created in VPOParoptTarget.cpp.

The one remaining alloca in the file "%loop.parameter.rec" was left as is
because a change to it cannot be tested as its test is currently XFAIL
pending some frontend changes.

We'll need to do similar changes for other constructs like omp for etc.
But changes like that will be more obtrusive as some code, like
the one emitting constructor/destructor calls, uses private allocas
as the insertion point for the calls. We'll need to move the allocas
while keeping those calls in place.

Change-Id: I2695fe298721a4b00359ab67738a3e185312ef65
JIRA: CMPLRLLVM-28118
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/325049
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>