[CMPLRLLVM-27552] [BPI] Add a new Heuristics for ADD(CMP, ADD(CMP,CMP)).
In spec_omp2012ref/372, the hottest BB's successor's frequency is not
accurate enough because there is three CMP conditions merged with AND/SEL
in the hottest BB in CFG simplify pass.
Actually the real taken branch's probabilty should be 12.5% which is much
smaller than 50%. Or else, the hottest BB's frequency and its variables'
weight is relatively small, so that its varialbles are easily evicted
and spilled into memory in RA phase.
The heuristic is the lowerest priority of all heuristics. Only if other
heuristic doesn't hint then start to check this heuristic.

Though cpu2017/549 has no regression in local SKX, it has -3% regression on
alloy. So adding condition to filt Fortran code. And adding loop restrict
too.

Change-Id: I0626bf2b8461a4b54646f9791f2e8b1a3dec079f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/326508
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Xiao, Wei3 <wei3.xiao@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>