Fix assert on the variable which is used in omp clause is not marked as used.
The problem only happens with constexpr variable, for constexpr variable,
variable is not marked during parser variable.   This is because compiler
might find some var's associate expressions may not actully an odr-used
later,  the variables get kept in MaybeODRUseExprs, in normal case, at
end of process fullExpr, the variable will be marked during the call to
CleanupVarDeclMarking(). Since we are processing expression of OpenMP
clauses, and the ActOnFinishFullExpr is not getting called that casue
variable is not get marked.

One way to fix this is to call CleanupVarDeclMarking() in EndOpenMPClause
for each omp directive.

This to fix https://bugs.llvm.org/show_bug.cgi?id=50206

Differential Revision: https://reviews.llvm.org/D101781

Change-Id: I3db9edf2185aea37aee115a493c3c4bfbbd46289
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327262
Reviewed-by: Yu, Jennifer <jennifer.yu@intel.com>
Tested-by: Yu, Jennifer <jennifer.yu@intel.com>