Pass the backedge taken count expr to isLoopEntryGuardedByCond() so it can be simplified using predicates guarding the loop.
Currently, simplification logic substitutes zext() by sext() when there
is a possibility of propagating sext() to the individual operands.

This change was motivated by the fact that a simplified expr helps loopopt
identify ZTT/loop entry guard. The change is guarded under HIR mode
(ScopedScalarEvolution) because I am seeing a couple of regressions in
the general case.

Change-Id: I058d42f6998e77ae7a9031e9cc30237f9807a0ba
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/325529
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>