[VPlan][SOA] Fix handling of scalar privates under SOA.
We want to treat scalar-private as SOA safe even if they are passed on
to a function. In CG, we would serialize the function calls and for that
we need the loop-private copy to be available in VPWidenMap and not the
first lane of VPScalarMap.

This change fixes the issue.

Change-Id: I2f2345beb4ccfb71038df51bf77bded19ac8c3d2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/325570
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>