CMPLRLLVM-27776: Add "mustprogress" loop and function attributes (part 1)
This change set introduces a new pass that will check if a simple loop
has bounds then it will be marked as "mustprogress". If all the loops
in a function were marked as "mustprogress", then the function will
be marked as "mustprogress".

The pass will be turned off by default for now.

Change-Id: I074d127fbd82fc02e98bbdb576a9657c0dce7250
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/326984
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>