CMPLRLLVM-28208: Match community behavior in visitAlloca()
Arvind was right, we should be matching the community behavior now in
visitAlloca() by returning "false" at the end of visitAlloca(). In the
past, we would return Base::visitAlloca(), but as the comment from
Arthur Eubanks states in Gerrit review 325637, that will always return
false, so it is better to return false. The actual detecting of whether
a dynamic alloca that matters has been seen will be noted by setting the
class variable HasDynamicAlloca. So, we should follow the community lead
also in the INTEL_CUSTOMIZATION code in this function as well.

Change-Id: I5c946c15716ff7acdcf8e78421daf97c6105ffba
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/325802
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Tested-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>