CMPLRLLVM-10459: Global dope vector constant propagation (part3)
This change set extends the validation of a dope vector to
consider that all store instructions for the dope vector
fields are executed if the alloc-site executes. This
guarantees that the data will be set when there is an
access to the array, else is an undefined behavior.

Also, this change set adds some extra details when printing
the dope vector info analysis for debugging purposes.

Change-Id: I463ccd0bb0acc6b0344ef40cfb853e03b6a85434
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/326325
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>