[XMAIN-REL] CMPLRLLVM-27805: Fix for regression in 511.povray and partial inlining
The current implementation of the simplified partial inlining looks
for one argument that will be used to access the exit block. There
is a chance that multiple arguments can branch into the exit block.
The analysis process can choose the wrong argument to check if
it is used to compute a loop, which could produce a false result.
This change set collects all the arguments that could branch into
and exit block and then identify which one will be used for
computing the loop.

This is a cherry pick from: https://git-amr-2.devtools.intel.com/gerrit/#/c/322599/

Change-Id: I69995b36d8cf4d41418c0ed8f86cf20cefed0ea9
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/326338
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>