[Aggressive Inlining] (CMPLRLLVM-27769): Skip unused functions for the analysis
Due to pulldown, there is a perf regression for 519.lbm on windows.
On windows, a dead comdat routine is not removed by earlier passes. The dead
routine is not marked with “norecurse”. Analysis stops as soon as it finds
a routine that is not marked “norecurse”. Analysis is unable to prove that
the dead routine is “norecurse”. The fix is to skip these unused dead routines
since Analysis runs under wholeprogramsafe.

Change-Id: Ib0f6041ed233c3243252898d190855a1e669caad
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/326773
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>