[NFC][VPlan][VPAA] Change VPAA interface to allow uncertainty in peeling selection
This patch is intended to make the difference between “no peel loop”
and “peeling is not guaranteed” more clear, as sometimes it seems to
be quite confusing. Also, the patch implements getGuaranteedPeeling()
function. The difference between PreferredPeeling and
GuaranteedPeeling is sometimes quite subtle. The former is supposed to
be used in cost model, while the latter is supposed to be used in
codegen where we need to compute conservative alignment.

VPlanAlignmentAnalysis::getAlignmentUnitStride() now accepts a pointer
to peeling variant, so that in case of uncertainty a nullptr can be
passed. If nullptr is passed, VPAA computes a conservative result
valid for any peeling.

CodeGen is changed accrodingly, so that it uses PreferredPeeling only
if it is guaranteed to be executed before vector loop.

Change-Id: Id4700bb6468c203fcb67e177c7f98be6d33eddfc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/326964
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>