[SYCL] Temporarily use host allocations for multi-device contexts
Currently Level Zero plugin allocates all buffers on a device with
index 0. If a context has multiple devices this allocation used on all
devices which results in an error. Temporarily make Level Zero plugin
to use host allocations for multi-device contexts because such
allocations are accessible by all devices. In case of a single discrete
device or if context consist of sub-devices of a same device we still
can use device allocation.

Change-Id: I3d553fc3d15266a57ad4b64e3a9720d5706192c8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327562
Reviewed-by: Gainullin, Artur <artur.gainullin@intel.com>
Tested-by: Gainullin, Artur <artur.gainullin@intel.com>