[Driver][FPGA] Improve output project folder behaviors for multi-file (#3697)
When performing multi-file compilations with AOT for FPGA, there is an
additional output report folder that is created with the separate aoc
call. This folder was continually over-written due to each subsequent
aoc call creating the same output folder name.

Add an additional capability to llvm-foreach:
let the user provide an --out-increment='file' argument. This will append
the number representing the current call made by llvm-foreach to the
given 'file' name.
Addresses:  CMPLRLLVM-24878

Change-Id: Id77747a859fa9d9d2ce2db328b3b43cc3c43e0a1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/326932
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327525
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>