[Driver][FPGA] Improve output project folder behaviors for multi-file (#3697)
When performing multi-file compilations with AOT for FPGA, there is an
additional output report folder that is created with the separate aoc
call. This folder was continually over-written due to each subsequent
aoc call creating the same output folder name.

Add an additional capability to llvm-foreach:
let the user provide an --out-increment='file' argument. This will append
the number representing the current call made by llvm-foreach to the
given 'file' name.

Change-Id: Id77747a859fa9d9d2ce2db328b3b43cc3c43e0a1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/326932
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chilamantula, Harini <harini.chilamantula@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>