Deduce original alloca base defined and used in def loop as dead after contraction.
Alloca bases which got contracted but had use in the original def loop were not
getting eliminated as dead.

This change eliminates such bases by proving that the def loop contains the
first definition of the alloca base in the function so it cannot be livein to
the loop.

It performs additional checks to make sure the def loop can be proven as dead
by checking for calls and any other defined bases. This would make it safe to
eliminate alloca stores thereby indicating to removeRedundantNodes() that the
entire loop can be eliminated.

Change-Id: I44cebb867193b5aa5e1b7c77b2b272d77757065f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/326958
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>