[VPlan][VLS] Add new checks into hir_vplan_opt_vls.ll test.
We want to check the number of groups created for the code from
hir_vplan_opt_vls.ll test. For this test VLSA is invoked mutiple times.
Each time the same set of memory references creates are new VLS group
which is a bug. This change set captures the current number of groups
created. Follow up change set in the compiler code will fix the number
of groups created.

Change-Id: I8e6c8689815e0e0a3e314b16dadbf9b199f109cd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327001
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>