[DTrans - OpaquePtr] Candidate selection for DeleteFieldOP pass
Adds the logic for selecting and pruning candidates that the
delete field transformation will be applied to. The same
heuristics that were used for the original delete fields
transformation are used here.

Much of this code is taken from the existing DeleteField pass,
but with changes needed to work with the new DTrans opaque
pointer infrastructure. The intent is that the original
delete field pass will be completely replaced by this at some
point, so there was not an attempt to break that into smaller
functions to share code between the new and old implementations.

Change-Id: Ifa02608804a313aca77abcc645448d3b8e38a33e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327573
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>