[1/2][VPlan][HIR] Add LIT test to reveal a bug in HIR vectorizer.
Currently VPlan HIR vectorizer codegen has a hack to initialize
conditional liveout temps that are not live-in with undef in loop
preheader. This can lead to incorrect results if the temp is
live-in to the parent Region. This patch adds a test to reveal
this bug in vectorizer.

Change-Id: Ideba0bca120291cd72ff126579017e53fc42831d
JIRA: CMPLRLLVM-27183
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327271
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>