[2/2][VPlan][HIR] Always create ExternalDef as CFG entry value of ambiguous symbases.
Currently VPlan HIR decomposer's fixPhiNodes pass uses ExternalDef
as CFG entry value for ambiguous Symbases only if they are live-in
to the candidate loop, it uses nullptr otherwise. This is not
entirely accurate since ambiguous temps that are not live-in to
the loop will be live-in at some parent loop level in the hierarchy.
We were incorrectly initializing such temps, which can also be
conditionally liveout, with undef before the vector loop which
revealed a bug in generated vector code. This patch resolves this
problem by updating decomposer to always have an ExternalDef for CFG
entry value of ambiguous symbases i.e. we don't rely on liveness to
initialize a temp at CFG entry. This change also enables us to get
rid of some temporary hacks that were added to processing liveouts
in both decomposer and CG.

JIRA : CMPLRLLVM-27183

Change-Id: Ibbec22b673e76f2864886577b06f93ba4bad927e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327272
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>