[VPlan][CM] Update SLP heuristics to check for reduction pattern that is better to SLP vectorize
The changes in this patch implement an interim fix to check and suppress
vectorization for a loop pattern that is better optimized by the SLP
vectorizer path. A proper fix will require generalized SLP cost modeling
in VPlan and/or also need several CM improvements that are currently being
worked on.

JIRA :: CMPLRLLVM-28045

Change-Id: I3a02e5a19ad8a4f6303827f61be98d94a9284d0b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327297
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>