[1/2][VPlan] Add test to reveal a bug in vectorizing sincos using vector-variants.
TTI does not use a strict ISA matching when choosing from available
set of vector variants for the matching process. This can lead to
a lower ISA variant being chosen for a higher target ISA (for example
SSE4.2 version being chosen AVX2 target). This is not illegal as long
as vector call is emitted in compliance to the ABI. This patch adds a
test to reveal a bug where this ABI compliance is missed for vector
sincos call.

JIRA : CMPLRLLVM-27551

Change-Id: I0f452a5ca3ef935247977df48318064fe0bc7254
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327727
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>